+++ Source/Tools/alive_api/alive_api.cpp
@@ -1,5 +1,5 @@
-#include "alive_api.hpp"
 #include "../AliveLibCommon/stdafx_common.h"
+#include "alive_api.hpp"
 #include "../AliveLibAE/Path.hpp"
 #include "../AliveLibAE/PathData.hpp"
 #include "../AliveLibAE/LvlArchive.hpp"
